+++ pyflo_bmi/__init__.py
@@ -1,4 +1,7 @@
-from .pyflo_model import *
+from .pyflo_bmi import *
 from .src.basins import *
+from .src.discrete_time_series import *
+from .src.conversions import *
 from .src.debug_utils import *
-from .bmi_pyflo import *
+from .pyflo_bmi_out import *
+from .unit_hydrograph_model import *
