+++ requirements.txt
@@ -1,3 +1,3 @@
 bmipy
+matplotlib
 numpy
-pyflo
